Refactor out generic functionality
